finished product

+++ db/migrate/20211215204906_create_ratings.rb
@@ -1,7 +1,7 @@
 class CreateRatings < ActiveRecord::Migration[6.1]
   def change
     create_table :ratings do |t|
-      t.integer :rating
+      t.float :rating
       t.integer :user_id
       t.integer :vendor_id
